+++ google/handlers/sac/index.ts
@@ -1,0 +1,10 @@
+import generateOauth from './generate-oauth';
+import { generateAccessToken } from './generate-token';
+
+/**
+ * Google Service Account Oauth Handlers
+ */
+export default {
+  generateOauth,
+  generateAccessToken,
+};