+++ google/helpers/index.ts
@@ -2,3 +2,4 @@ export { default as api } from './API';
 export { default as resolveToken } from './resolve-token';
 export { default as stringizeScope } from './stringize-scope';
 export { default as googleApiRequest } from './requester';
+export { default as createJwtToken } from './signJwt';